Moves things around: * Classes out of FilterAdapter * Functionality from the FilterAnimator * Unused methods from FilterAdapter are removed
