+++ core/src/main/java/io/plaidapp/core/ui/filter/FilterHolderInfo.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2019 Google, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.plaidapp.core.ui.filter
+
+import androidx.recyclerview.widget.RecyclerView
+
+class FilterHolderInfo : RecyclerView.ItemAnimator.ItemHolderInfo() {
+
+    var doEnable: Boolean = false
+    var doDisable: Boolean = false
+    var doHighlight: Boolean = false
+
+    companion object {
+
+        const val FILTER_ENABLED = 1
+        const val FILTER_DISABLED = 2
+        const val HIGHLIGHT = 3
+    }
+}
